Refactor reading status and control dbs.
Add a `read_db()` function that takes an array of lines from a db of the
form the `status` db is in, or a package `CONTROL` file, which is in the
same format, and returns an array of ordered hashes.

Use this function instead of having duplicated code to make the ordered
hashes in three places.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>