Add Port-Version to package zips,remove -nonewline
Add an -r$port_version suffix to the version parts of the created
package zip files, if there is a Port-Version set in the status file.

Also remove -nonewline from terminal status messages, because this does
not work from cmake, the first part of the message does not show up
until the second part is printed. So to indicate progress, print the
status message and the "done." part on separate lines.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>