Fixes for Linux
Filter out package depends on themselves, e.g. fribidi on Linux depends
on itself. This fixes the infinite recursion this causes in the
dependency calculator.

Support build dependencies that are unqualified with a triplet, guess
the host triplet in these cases.

Use `ni -it dir` instead of `mkdir`, because `mkdir` on Linux is the
Linux command and behaves differently.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>