Implement vcpkg-list
Implement a vcpkg-list command that works almost the same way as `vcpkg
list` because the default VS version of vcpkg does not support this
command, and for the time being I need it for my cmake code.

TODO: the sort order in my version is slightly off from `vcpkg list`,
that will need to be fixed later.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>